requirements.txt
streamlit
pandas
python-dateutil
icalendar
pytz
pdfplumber
unidecode
openpyxl
